# pkg/storage: Querying flat profiles