--- a/pkg/symbolizer/symbolizer_test.go
+++ b/pkg/symbolizer/symbolizer_test.go
@@ -445,6 +445,7 @@ func setup(t *testing.T) (*grpc.ClientConn, *debuginfo.Store, TestProfileMetaSto
 		trace.NewNoopTracerProvider().Tracer(""),
 		db,
 		mStr,
+		true,
 	)
 
 	lis, err := net.Listen("tcp", ":0")