… pkg/query/query_test.go
… pkg/storage/benchmark/db-appends.txt
… pkg/storage/benchmark/db-iterator.txt
… pkg/storage/benchmark_test.go
… pkg/storage/db_test.go
… pkg/storage/flamegraph_flat_test.go
… pkg/storage/profile_flat_test.go
… pkg/storage/querier_bench_test.go
… pkg/storage/series_iterator.go
… pkg/storage/series_iterator_merge_test.go
… pkg/storage/series_iterator_range.go
… pkg/storage/series_iterator_range_test.go
… pkg/storage/series_iterator_root.go
… pkg/storage/series_iterator_root_test.go
… pkg/storage/series_test.go