Implement querying the new, more simple, `[]*Sample` from the QueryRange and Query.singleProfile handlers.

For now, we pass a `profileTrees` bool into some structs to decide whether or not to use the new storage or old storage way of representing trees. Eventually, we'll clean it up by removing the profileTree entirely.