@metalmatze 2021-11-16
Benchmarks are looking good. And I already know what to do next for appends to get this even lower than the previous benchmarks.

appends:
```
name        old time/op    new time/op    delta
Appends-24    12.5ms ± 3%    16.8ms ± 1%   ~     (p=0.100 n=3+3)

name        old alloc/op   new alloc/op   delta
Appends-24    6.26MB ± 0%    9.84MB ± 0%   ~     (p=0.100 n=3+3)

name        old allocs/op  new allocs/op  delta
Appends-24      122k ± 0%      140k ± 0%   ~     (p=0.100 n=3+3)
```
iterators: (look at this! :tada: :rocket:)
```
name         old time/op    new time/op    delta
Iterator-24     104µs ± 5%      62µs ± 1%   ~     (p=0.100 n=3+3)

name         old alloc/op   new alloc/op   delta
Iterator-24    28.0kB ± 0%     0.6kB ± 0%   ~     (p=0.100 n=3+3)

name         old allocs/op  new allocs/op  delta
Iterator-24     1.70k ± 0%     0.00k ± 0%   ~     (p=0.100 n=3+3)
```