@brancz 2021-11-17
Do we have an idea why appends are slower? I would have expected that due to skipping trees we should be quite a bit faster?