@metalmatze 2021-11-17
Most of the time in appends is spent in `makeStacktraceKey` and that's exactly what we're going to tackle next:
https://share.polarsignals.com/9aa74a3/

Most significantly, we didn't update the benchmarks after introducing UUIDs. So what we're really comparing above is profileTrees without UUIDs to the new flat samples with UUIDs too.
Let me provide you with a fairer comparison in a minute.