@metalmatze 2021-11-17
tree with UUIDs vs flat with UUIDs
```
name        old time/op    new time/op    delta
Appends-24    14.3ms ± 2%    16.8ms ± 1%   ~     (p=0.100 n=3+3)

name        old alloc/op   new alloc/op   delta
Appends-24    8.98MB ± 0%    9.84MB ± 0%   ~     (p=0.100 n=3+3)

name        old allocs/op  new allocs/op  delta
Appends-24      122k ± 0%      140k ± 0%   ~     (p=0.100 n=3+3)
```